Updated; build key-exchange, node-consumer
2022.09.06 - isanghyeon

+++ core/node-consumer/fpe/build.sh
@@ -5,6 +5,11 @@ source ./.env
 echo "[*] Environment :: " "$ENV"
 echo "[*] Path        :: " "$Path"
 echo "[*] Build       :: " "$TIMESTAMP-$ENV"
+check="N"
+
+sleep 3
+
+apt-get install -y docker-compose
 
 sleep 3
 
@@ -12,15 +17,24 @@ docker-compose stop
 docker-compose down
 docker-compose rm -f
 
-docker rmi $Image:$TIMESTAMP-$ENV
+sleep 3
 
-docker build -t $Image:$TIMESTAMP-$ENV $Path
+read -p "Delete legacy environment or docker image? [y/N] " check
 
-if [ $? -eq 1 ]; then
-  echo "[-] Docker image build failed..."
-else
-  echo "[+] Docker image build completed..."
+sleep 3
+
+if [ ${check^^} == "Y" ]; then
+    docker rmi $Image:$TIMESTAMP-$ENV
+    docker build -t $Image:$TIMESTAMP-$ENV $Path
+
+    if [ $? -eq 1 ]; then
+        echo "[-] Docker image build failed..."
+    else
+        echo "[+] Docker image build completed..."
+    fi
 fi
+
+echo "[+] Docker image build completed..."
 
 sleep 3
